Update class.yith-wcwl.php
While validating a page that has woocommerce with wishlist plugin enabled I was getting a '“& did not start a character reference' error. So after looking into files I found an escaped ampersand character in the '$url' variable.

On line 283 originally it was:

 $url = YITH_WCWL_URL . "yith-wcwl-ajax.php?action=add_to_wishlist&add_to_wishlist=" . $product->id;

After changing it to

 $url = YITH_WCWL_URL . "yith-wcwl-ajax.php?action=add_to_wishlist&amp;add_to_wishlist=" . $product->id;

The error was gone. So that should be changed in next version.